norm and denorm image
